### 2016-05-17 #23 @whedon
Submission: Scowl: a Scala DSL for programming with the OWL API
**Submitting author:** @balhoff (<a href="http://orcid.org/0000-0002-8688-6599">James P. Balhoff</a>)
**Repository:** <a href="https://github.com/phenoscape/scowl" target ="_blank">https://github.com/phenoscape/scowl</a>
**Version:** v1.1
**Archive:** <a href="http://dx.doi.org/10.5281/zenodo.51556" target ="_blank">http://dx.doi.org/10.5281/zenodo.51556</a>

- [x] **Archive:** Does the <a target="_blank" href="http://dx.doi.org/10.5281/zenodo.51556">software archive</a> resolve?

- [x] **Performance:** Have the performance claims of the software been confirmed?

- [x] **Functionality documentation:** Is the functionality of the software documented to a satisfactory level (e.g. API method documentation)?

Compiled paper PDF: [10.21105.joss.00023.pdf](https://github.com/openjournals/joss-reviews/files/269284/10.21105.joss.00023.pdf)
