### 2016-05-28 #23 @xuanxu
Review finished, every point is checked but one:

### Community guidelines
There is no mention on how to contribute, report bugs or get support.

Please add a 'How to contribute' section to the README with at least a brief statement encouraging posible contributors to help pointing them to open a new github issue. Add also information on how to report a problem (opening an issue should be OK) and seek support (ideas: mailing list, contact email, github issue with a specific tag).
Even better would be to include a CONTRIBUTING.md file with detailed instructions.

### Recommendation:
I recommend publication after some minor revisions: contributing guidelines should be added to the documentation.
