### 2016-05-31 c607107
Added CONTRIBUTING, info about support, updated paper references.
