+++ paper.bib
@@ -1,13 +1,13 @@
 % Generated by Paperpile. Check out http://paperpile.com for more information.
 % BibTeX export options can be customized via Settings -> BibTeX.
 
-@MISC{OWLAPI,
+@ONLINE{OWLAPI,
   title        = "The {OWL} {API}",
   url          = "http://owlapi.sourceforge.net/",
   howpublished = "\url{http://owlapi.sourceforge.net/}"
 }
 
-@MISC{Horridge2009,
+@ONLINE{Horridge2009,
   title        = "{OWL} 2 web ontology language manchester syntax",
   author       = "Horridge, Matthew and Patel-Schneider, Peter F",
   year         =  2009,
@@ -15,7 +15,7 @@
   howpublished = "\url{https://www.w3.org/TR/owl2-manchester-syntax/}"
 }
 
-@MISC{Motik2009,
+@ONLINE{Motik2009,
   title        = "{OWL} 2 web ontology language: Structural specification and
                   functional-style syntax",
   author       = "Motik, Boris and Patel-Schneider, Peter F and Parsia, Bijan
@@ -27,7 +27,7 @@
   howpublished = "\url{https://www.w3.org/TR/owl2-syntax/}"
 }
 
-@MISC{OWL2009,
+@ONLINE{OWL2009,
   title        = "{OWL} 2 Web Ontology Language Document Overview",
   author       = "{W3C OWL Working Group}",
   month        =  "27~" # oct,
