+++ CONTRIBUTING.md
@@ -1,0 +1,7 @@
+# Contributing to Scowl
+
+## Have a question or problem?
+If you have questions about how to use Scowl, feel free to send an email to balhoff@gmail.com, or [open an issue on the tracker](https://github.com/phenoscape/scowl/issues).
+
+## Found an issue?
+If you find a bug in Scowl, please [open an issue on GitHub](https://github.com/phenoscape/scowl/issues). If you would like to create a fix yourself, pull requests are welcome.
