### 2016-05-31 #23 @balhoff
Thanks @xuanxu! I updated the [README](https://github.com/phenoscape/scowl#question-or-problem) and added [CONTRIBUTING.md](https://github.com/phenoscape/scowl/blob/master/CONTRIBUTING.md). Please let me know what you think.

I also updated [paper.bib](https://github.com/phenoscape/scowl/blob/master/paper.bib), hoping to eliminate duplicate URLs in the reference output.
