### 2016-05-31 4b505e4
Added JOSS badge.
