### 2016-05-31 #23 @xuanxu
Thanks @balhoff! It looks good to me.
All points checked, this is publication-ready @arfon.
