### 2016-06-01 #23 @balhoff
Great! @arfon would it be possible to regenerate the PDF at this point? I had made some tweaks to the references file hoping to remove duplicated URLs from the output.
