### 2016-06-01 #23 @arfon
@balhoff - I regenerated it when I generate the DOI: https://github.com/openjournals/joss-papers/blob/master/joss.00023/10.21105.joss.00023.pdf

Does ☝️ look good? If not, we can update it.
