### 2016-06-01 #23 @balhoff
@arfon I just updated paper.bib to remove the `howpublished` field. The website citations were showing the URL twice. Hopefully this will fix it. Can you regenerate? Thanks!
