### 2016-06-01 #23 @arfon
Is this better? https://github.com/openjournals/joss-papers/blob/master/joss.00023/10.21105.joss.00023.pdf
